// lib/platform/types.ts
export interface Ticket {
  ticket: string;
  randstr?: string;
  provider: string;
  createdAt: number;
}
// ...
export interface ITicketProvider {
  readonly providerId: string;
  produce(): Promise<Ticket | null>;
  startBatch(): void;
  stopBatch(): void;
  destroyActive(): void;
}